int and other cursor changes

+++ src/main/java/com/esri/core/geometry/OperatorExportToGeoJsonCursor.java
@@ -69,7 +69,7 @@ public class OperatorExportToGeoJsonCursor extends StringCursor {
 	}
 
 	@Override
-	public long getID() {
+	public int getID() {
 		return m_geometryCursor.getGeometryID();
 	}
